+++ tests/testthat.R
@@ -1,4 +1,0 @@
-library(testthat)
-library(RSimpleFFI)
-
-test_check("RSimpleFFI")